% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-fncs-v19.R
\name{add.dY}
\alias{add.dY}
\title{Simple helper function to add elements dY, dY.in and dY.out to an existing list with the same elements}
\usage{
add.dY(A = NULL, fn = NULL, include = F)
}
\arguments{
\item{A}{a list with elements dY, dY.in and dY.out}

\item{fn}{a function that produces a list with elements dY, dY.in and dY.out}
}
\value{
the modified list A after adding the result of fn to elements dY, dY.in and dY.out
}
\description{
Simple helper function to add elements dY, dY.in and dY.out to an existing list with the same elements
}
