% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{add.vaccination.results}
\alias{add.vaccination.results}
\title{Adds vaccination results in case vaccination is once per year ; these data will be part of the output}
\usage{
add.vaccination.results(t, p, dvacc, rownamesZ, subject = "VXa")
}
\arguments{
\item{t}{current time}

\item{p}{an environment with initialized model parameters}

\item{dvacc}{an environment with initialized model parameters}

\item{rownamesZ}{an environment with initialized model parameters}

\item{subject}{refers to one of the VXa incidence files e.g. VXaXi_1 etc}
}
\value{
updated p$vaccinated
}
\description{
Adds vaccination results in case vaccination is once per year ; these data will be part of the output
}
