% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-age-related-fncs-v14.R
\name{adjust.age.groups}
\alias{adjust.age.groups}
\title{adjust age groups of incidence file to age groups defined in params environment}
\usage{
adjust.age.groups(
  p = NULL,
  M = NULL,
  compress.1st = F,
  proportions = F,
  aggregate = T
)
}
\arguments{
\item{p}{params environment}

\item{M}{matrix with lower limit of age group in column names}

\item{compress.1st}{logical to indicate if incidence in age groups < 5 should be compressed in 1st i.e. death at birth}

\item{proportions}{logical to indicate if numbers per row are proportions (and should add up to 1.0)}

\item{aggregate}{logical to indicate if single year age groups should be aggregated to p$AGES}
}
\value{
a matrix with age groups according to p$AGES
}
\description{
initializes VXa, SES, RISK, HIV, or TB progression from the XML specification
}
