% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{aging.and.births.and.update.contact.matrix.event}
\alias{aging.and.births.and.update.contact.matrix.event}
\title{Handles aging, adding newborns, adjusting population composition, updating the contacts matrices}
\usage{
aging.and.births.and.update.contact.matrix.event(t, X, p)
}
\arguments{
\item{t}{current time}

\item{X}{a \code{double[nVXaSESRISKHIVTB x nAGES]} vector of state variables}

\item{p}{an environment with initialized model parameters}
}
\value{
a \code{double[nVXaSESRISKHIVTB x nAGES]} vector of adjusted state variables
}
\description{
\code{aging.and.births.and.update.contact.matrix.event()} is called once a year, at the start of a new year
}
\details{
aging.and.births.and.update.contact.matrix.event() takes care of:
\describe{
\item{\code{aging}:}{moving (part of) the contents of the state matrix \code{Y} (after converting the vector \code{X}) to the next higher age group }
\item{\code{births}:}{adding newborns}
\item{\code{fiting demography}:}{adjusting the population composition to UNPOP data (baseline) or reusing recorded population adjustments (intervention)}
\item{\code{rescaling}:}{rescaling the population (usually in 1950)}
\item{\code{updating the contacts matrices}:}{i.e. balancing the contacts matrix to ensure equal number of contacts between ego - alter age groups and vice versa which is necessary if population composition changes}
\item{\code{yearly incidence}:}{adding incident cases if incidence is defined as a yearly event rather than a rate}
}
}
