% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-matrix-fncs-v12.R
\name{calc.all.indices.for.dim}
\alias{calc.all.indices.for.dim}
\title{Calculates the indices (e.g. 1..10 for HIV) of a specific dimension for all stages of the dimension}
\usage{
calc.all.indices.for.dim(p, dimname)
}
\arguments{
\item{p}{an environment with initialized model parameters}

\item{dimname}{a string with the name of the dimension e.g. "VXa" or "HIV"}
}
\value{
a vector of length p$nVXaSESRISKHIVTB with indices within the specified dimension
}
\description{
Calculates the indices (e.g. 1..10 for HIV) of a specific dimension for all stages of the dimension
}
