% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{calc.birthrate}
\alias{calc.birthrate}
\title{Creates an approxfun of birthrate as a function of time from a filename and countrycode}
\usage{
calc.birthrate(fname = NA, countrycode = NA)
}
\arguments{
\item{fname}{filename}

\item{countrycode}{3 letter ISO country code (e.g. 'ZAF')}
}
\value{
an approxfun birthrate(t)
}
\description{
Creates an approxfun of birthrate as a function of time from a filename and countrycode
}
