% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-query-fncs-v5.R
\name{calc.count}
\alias{calc.count}
\title{helper function called from calc.targets() to calculate a count value}
\usage{
calc.count(z, target, p = NULL)
}
\arguments{
\item{z}{a data frame with stocks output}

\item{target}{a subset of the columns of a single row of the targets defining either the numerator or the denominator to be calculated}

\item{p}{initialized model parameters}
}
\value{
a single number i.e. the result of querying the stocks data frame with a query generated from the target parameter
}
\description{
helper function called from calc.targets() to calculate a count value
}
