% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-query-fncs-v5.R
\name{calc.flow}
\alias{calc.flow}
\title{helper function called from calc.targets() to calculate a flow value}
\usage{
calc.flow(z, target, p = NULL)
}
\arguments{
\item{z}{a data frame with flows output}

\item{target}{a subset of the columns of a single row of the targets defining either the numerator or the denominator to be calculated}

\item{p}{initialized model parameters}
}
\value{
a single number i.e. the result of querying the flows data frame with a query generated from the target parameter to produce either inflow or outflow
}
\description{
helper function called from calc.targets() to calculate a flow value
}
