% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-matrix-fncs-v12.R
\name{calc.names.for.dim}
\alias{calc.names.for.dim}
\title{Pulls out the bare names of the stages of a specific dimension from p$VXaSESRISKHIVTB}
\usage{
calc.names.for.dim(p, dimname)
}
\arguments{
\item{p}{an environment with initialized model parameters}

\item{dimname}{a string with the name of the dimension e.g. "VXa" or "HIV"}
}
\value{
a vector of length p$nVXaSESRISKHIVTB with names of the stages of the specified dimension
}
\description{
Pulls out the bare names of the stages of a specific dimension from p$VXaSESRISKHIVTB
}
