% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-matrix-fncs-v12.R
\name{calc.small.matrix.indices.for.one.full.dim.and.specific.single.dim.indices.from.list}
\alias{calc.small.matrix.indices.for.one.full.dim.and.specific.single.dim.indices.from.list}
\title{Creates the indices for one full dimension and a subset of single other dimensions}
\usage{
calc.small.matrix.indices.for.one.full.dim.and.specific.single.dim.indices.from.list(
  p,
  list
)
}
\arguments{
\item{p}{an environment with initialized model parameters}

\item{a}{named list of dimensions all of length 1 except one of full length}
}
\value{
the indices of a subset of p$VXaSESRISKHIVTB to be used to create a small transition / progression matrix
}
\description{
Called from create.full.parameter.matrix.for.age.group()
}
