% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-query-fncs-v5.R
\name{calc.targets}
\alias{calc.targets}
\title{calculates target model values from the output generated by running the model}
\usage{
calc.targets(z, targets, p = NULL)
}
\arguments{
\item{z}{the output from running the model: a list with elements stocks and flows}

\item{targets}{a data frame with targets (from reading the targets csv)}

\item{p}{initialized model parameters}
}
\value{
the targets data frame with a new column 'model' with values calculated from the model output
}
\description{
calculates target model values from the output generated by running the model
}
