% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{calc.vac.flows}
\alias{calc.vac.flows}
\title{Calculates the vaccination flows from p$vaccinated (the yearly update of once per year vaccinations)}
\usage{
calc.vac.flows(p)
}
\arguments{
\item{p}{an environment with initialized model parameters}
}
\value{
a data frame of yearly vaccination flows (if option once.per.year=T) to be included in model output e.g. as result$vaccinated
}
\description{
Calculates the vaccination flows from p$vaccinated (the yearly update of once per year vaccinations)
}
