% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{check.names.values}
\alias{check.names.values}
\title{checks that all XML parameters at a specified xmlpath have a name and a value attribute}
\usage{
check.names.values(xmlpath, xmlparams)
}
\arguments{
\item{xmlpath}{the path to the XML parameters}

\item{xmlparams}{a node list of XML parameters}
}
\description{
checks that all XML parameters at a specified xmlpath have a name and a value attribute
}
