% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{check.sum.to.1}
\alias{check.sum.to.1}
\title{Checks that the sum of specified fractions equals 1.0}
\usage{
check.sum.to.1(v)
}
\arguments{
\item{v}{a double[] vector}
}
\value{
a vector with fractions that sum to 1.0
}
\description{
If all zero, replaces first element by 1.0
Generates an error if not all zero or sum != 1.0
}
