% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{col.names.from.schema}
\alias{col.names.from.schema}
\title{returns all possible attribute names for a specific XML parameter (e.g. "TB.parameter") from the XML Schema}
\usage{
col.names.from.schema(xmlschema, xmlparameter)
}
\arguments{
\item{xmlschema}{the XML schema}

\item{xmlparameter}{the XML parameter}
}
\value{
a character vector of attribute names e.g. "VXa.stage" "SES.stage" etc for xmlparameter "TB.parameter"
}
\description{
returns all possible attribute names for a specific XML parameter (e.g. "TB.parameter") from the XML Schema
}
