% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{constant.parameters}
\alias{constant.parameters}
\title{Selects constant parameters from a data frame}
\usage{
constant.parameters(parameters.df = NULL)
}
\arguments{
\item{parameters.df}{a data frame with parameters}
}
\value{
a data frame with constant parameters
}
\description{
Selects constant parameters from a data frame
}
