% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-age-related-fncs-v14.R
\name{create.aging.matrix}
\alias{create.aging.matrix}
\title{Creates a matrix to do aging once per year (see Schenzle et al.) ; called from function aging.and.births.and.update.contact.matrix.event()}
\usage{
create.aging.matrix(p)
}
\arguments{
\item{p}{parameters environment}
}
\value{
aging matrix
}
\description{
Creates a matrix to do aging once per year (see Schenzle et al.) ; called from function aging.and.births.and.update.contact.matrix.event()
}
