% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{create.approx.fun}
\alias{create.approx.fun}
\title{Creates an approximate function from an XML element and a data frame with parameter values}
\usage{
create.approx.fun(row, df = NULL)
}
\arguments{
\item{row}{a  named vector with names  'parameter', 'times', 'values' that will be parsed}

\item{df}{a  data frame with parameter values}
}
\value{
an approxfun f(t)
}
\description{
For example: 
<multiplier name="etamul" times="1960:2020" values="1/(1+exp(-sh*(x-midx)))"/>
}
