% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{create.approx.func.from.rownames}
\alias{create.approx.func.from.rownames}
\title{Creates an approximate function to index a data frame using years specified in the rownames}
\usage{
create.approx.func.from.rownames(df, offset = 0)
}
\arguments{
\item{df}{a data frame with row names that can be converted to numeric}

\item{offset}{an optional offset}
}
\value{
an approxfun to generate a row index from a row name (usually a year)
}
\description{
Creates an approximate function to index a data frame using years specified in the rownames
}
