% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{create.contacts.matrices}
\alias{create.contacts.matrices}
\title{Creates contacts matrices}
\usage{
create.contacts.matrices(paths = NULL, model.params = NULL)
}
\arguments{
\item{paths}{paths to various files ; see set.paths() function}

\item{model.params}{params environment (to be updated)}
}
\value{
a list(defined,M,MplustM) i.e. a logical to indicate if contacts matrices are defined and if so 
non NULL M and MplustM (i.e. M + t(M))
}
\description{
See also the init.incidence() function in TBVx-data-reading-fncs-v##.R
}
