% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-data-reading-fncs-v13.R
\name{create.empty.population}
\alias{create.empty.population}
\title{Creates a matrix with an empty population}
\usage{
create.empty.population(rownms, colnms)
}
\arguments{
\item{rownms}{rownames}

\item{colnms}{colnames}
}
\value{
a matrix with 0's in rows and columns
}
\description{
Creates a matrix with an empty population
}
