% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-data-reading-fncs-v13.R
\name{create.filename}
\alias{create.filename}
\title{Creates a file name based on the passed parameters, a hash of model parameters and the date and time}
\usage{
create.filename(
  path = NULL,
  model.params = NULL,
  runnr = NA,
  runtype = NA,
  score = NA
)
}
\arguments{
\item{path}{the path to which the file will be written}

\item{model.params}{environment with initialized model parameters}

\item{runnr}{runnr (used when running the optimizer)}

\item{runtype}{string which specifies the runtype e.g. "flows" or "stocks" etc}

\item{score}{score (used when running the optimizer)}
}
\value{
a file name that is usually unique due to including a hash of the model parameters and the date and time
}
\description{
Creates a file name based on the passed parameters, a hash of model parameters and the date and time
}
