% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{create.full.parameter.matrix.for.age.group}
\alias{create.full.parameter.matrix.for.age.group}
\title{creates a parameter matrix for an age group e.g. for progression, infectivity, etc}
\usage{
create.full.parameter.matrix.for.age.group(
  p = NULL,
  ag = NULL,
  params.df,
  transitions,
  auxparameters,
  values,
  dim.name
)
}
\arguments{
\item{p}{initialized parameters environment}

\item{ag}{age group}

\item{params.df}{a data frame of parameter values (by name and value) including dependencies (for all dimensions and age groups except current dimension)}

\item{transitions}{a data frame of transition with columns from, to and rate}

\item{auxparameters}{auxiliary parameters defined in the transition or treatment matrix (usually not used)}

\item{values}{a logical to indicate if values should be read for stages (e.g. for TB infectivity) or for transitions (from, to)}

\item{dim.name}{the dimension e.g. "TB"}
}
\value{
a double p$DIMLENGTHS x p$DIMLENGTHS matrix with parameter values to be used to calculate the contribution of e.g. progression to dY/dt
}
\description{
calls  create.small.matrix.from.query()
}
