% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-matrix-fncs-v12.R
\name{create.incidence.matrix}
\alias{create.incidence.matrix}
\title{Creates an incidence matrix (with -1 to indicate 'from' state and +1 to indicate 'to' states)}
\usage{
create.incidence.matrix(p, dim, from, to, sel = NULL)
}
\arguments{
\item{p}{an environment with initialized model parameters}

\item{dim}{a string with the name of the dimension e.g. "VXa" or "HIV"}

\item{from}{the name of the from state e.g. "HIV-"}

\item{to}{the name(s) of the to state(s) e.g. "HIVu1" and "HIVcount"}
}
\value{
a double p$nVXaSESRISKHIVTB x p$nVXaSESRISKHIVTB matrix to be used to implement state transitions due to (VXa, HIV) incidence just with -1 and +1 in the appropriate positions
}
\description{
Creates an incidence matrix (with -1 to indicate 'from' state and +1 to indicate 'to' states)
}
