% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{create.list.of.parameter.matrices.by.age.group}
\alias{create.list.of.parameter.matrices.by.age.group}
\title{creates a list of parameter matrices by age group e.g. for progression, infectivity, etc}
\usage{
create.list.of.parameter.matrices.by.age.group(
  p,
  age.groups,
  fixed.parameters,
  xmlpath,
  xmltransitions,
  xmlauxparameters,
  values,
  col.names = NULL,
  dim.name = NULL
)
}
\arguments{
\item{p}{initialized parameters environment}

\item{age.groups}{}

\item{fixed.parameters}{an environment with xmlpath, xmlparameter, dim.name, col.names, age.groups, age.ranges, parameter.matrices, timed.parameters}

\item{xmlpath}{a node list of XML parameters defined in the transition matrix (if any) i.e. "//TB/TB.transmission/transition.matrix/parameter"}

\item{xmltransitions}{a node list of XML timed parameters i.e. time dependent multipliers from e.g. "//TB/TB.transmission/transition.matrix/multiplier"}

\item{xmlauxparameters}{a logical to indicate if values should be read for stages (e.g. for TB infectivity) instead of for transitions (from, to)}

\item{values}{a logical to indicate if values should be read for stages (e.g. for TB infectivity) or for transitions (from, to)}

\item{col.names}{the column names of the transition matrix}

\item{dim.name}{the dimension e.g. "TB"}
}
\value{
a list (by age group) of parameter matrices for the specified dimension
}
\description{
calls  create.full.parameter.matrix.for.age.group()
}
