% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{create.small.matrix.from.query}
\alias{create.small.matrix.from.query}
\title{creates a small parameter matrix for an age group, a dimension (e.g. "TB") and a selection of index values of the other dimensions from a query}
\usage{
create.small.matrix.from.query(
  parms_,
  query_ = NA,
  z_,
  m_,
  auxparameters_,
  values_,
  dim.name_
)
}
\arguments{
\item{parms_}{initialized parameters environment}

\item{query_}{the query generated by query.from.col.names()}

\item{z_}{a data frame of parameter values (by name and value) including dependencies (for all dimensions and age groups except current dimension)
from which a subset will be generated using the query}

\item{m_}{a data frame with transition rates (if values=F) or values by stage (if values=T e.g. for infectivity)}

\item{auxparameters_}{auxiliary parameters defined in the transition or treatment matrix (usually not used)}

\item{values_}{a logical to indicate if values should be read for stages (e.g. for TB infectivity) or for transitions (from, to)}

\item{dim.name_}{the dimension e.g. "TB"}
}
\value{
a double p$DIMLENGTHS x p$DIMLENGTHS matrix with parameter values to be used to calculate the contribution of e.g. progression to dY/dt
}
\description{
creates a small parameter matrix for an age group, a dimension (e.g. "TB") and a selection of index values of the other dimensions from a query
}
