% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-data-reading-fncs-v13.R
\name{deathrate.from.data}
\alias{deathrate.from.data}
\title{Reads a csv file with death rates(from UNPOP) for a country by year (in rows) and single age years (in columns)}
\usage{
deathrate.from.data(paths = NULL, model.params = NULL)
}
\arguments{
\item{paths}{the path to which the country code is appended file name}

\item{model.params}{an environment with parameters that is used to get the file name and country code from}
}
\value{
a matrix with the death rates by year (rownames) and age group (colnames)
}
\description{
Reads a csv file with death rates(from UNPOP) for a country by year (in rows) and single age years (in columns)
}
