% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{default.value.from.schema}
\alias{default.value.from.schema}
\title{retrieves the default value of an XML attribute from the XML Schema}
\usage{
default.value.from.schema(xmlschema, attrname, type)
}
\arguments{
\item{xmlschema}{the XML schema}

\item{attrname}{the attribute name}

\item{type}{the type of the attribute (e.g. double, fraction, integer, ...)}
}
\value{
the default value of the attribute passed as an argument
}
\description{
retrieves the default value of an XML attribute from the XML Schema
}
