% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{derivs.deSolve}
\alias{derivs.deSolve}
\title{Calculate dY/dt (a vector of derivates of state variables)}
\usage{
derivs.deSolve(t, X, p)
}
\arguments{
\item{t}{current time}

\item{X}{a double[] vector of state variables}

\item{p}{an environment with initialized model parameters}
}
\value{
a \code{list[2]} with two \code{double[]} vectors: \code{dY} and \code{dY.HIVTBdeaths}
}
\description{
\code{derivs.deSolve()} calculates \code{dY/dt} [a vector of derivatives of state variables] at timepoint \code{t} using parameters \code{p}
}
\details{
Further details:

\describe{
\item{\code{Y = Y.matrix.from.X.vector(X,p)}}{convert vector X to matrix Y}
\item{\code{dY = matmul.by.age.group(p$TBHIVptr,Y)}}{Calculate dY due to TB and HIV progression and treatment [only for constant matrices] ; \code{TBHIVptr} has at least \code{TBp} and is optionally combined with \code{HIVp} and \code{TBtr} and \code{HIVtr}}
\item{\code{dY = dY + derivs.Tm(t,Y,p)}}{TB transmission}
\item{\code{if (!is.null(p$TBtr)) \emph{update dY}}}{ with TB treatment (if not already in TBHIVptr)}
\item{\code{for (i in seq_along(p$inci$TB))\emph{update dY}}}{ with TB incidence}
\item{\code{if (p$nHIV>1)\emph{update dY}}}{ with HIV: progression, treatment, incidence (except when already in HIVTBptr)}
\item{\code{if (p$nRISK>1 & !is.null(p$RISKp))\emph{update dY}}}{with RISK progression}
\item{\code{if (p$nSES>1  & !is.null(p$SESp))\emph{update dY}}}{with SES progression}
\item{\code{if (p$nVXa>1)\emph{update dY}}}{with VXa: progression, incidence}
}
}
\section{A special section}{


Text for section A
}

