% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{derivs.inci.common}
\alias{derivs.inci.common}
\title{Calculates the derivative dY/dt of matrix Y due to incidence}
\usage{
derivs.inci.common(t = NA, Y = NA, p = NA, p.inci = NA)
}
\arguments{
\item{t}{current time}

\item{Y}{a \code{double[nVXaSESRISKHIVTB x nAGES] matrix} of state variables}

\item{p}{an environment with initialized model parameters}

\item{p.inci}{an environment derived from options in the XML e.g.

\preformatted{
<HIV.incidence>
  <incidence.data file="data/HIV-incidence.txt" times="1980,2051" values="lambdaH,lambdaH" proportions="false" denominator="susc"/>
</HIV.incidence>
}

and from the file that defines the incidence (e.g. “data/HIV-incidence.txt”)

The most important data structure is \code{p.inci$inci.matrix} which is based on the incidence data file.}
}
\value{
a \code{double [nVXaSESRISKHIVTB x nAGES] matrix} of derivatives of state variables
}
\description{
\code{derivs.inci.common()} calculates the contribution of Mtb transmission to the derivatives of \code{Y} at time \code{t} with parameters \code{parms}
}
