% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{derivs.inci.in.out}
\alias{derivs.inci.in.out}
\title{Calculates the derivative dY/dt of matrix Y due to incidence split into negative and positive contributions (outflow and inflow)}
\usage{
derivs.inci.in.out(t = NA, Y = NA, p = NA, p.inci = NA, once = F)
}
\arguments{
\item{t}{current time}

\item{Y}{a double[nVXaSESRISKHIVTB x nAGES] matrix of state variables}

\item{p}{an environment with initialized model parameters}

\item{p.inci}{incidence data structure including incidence matrices etc}

\item{once}{logical that should match p.inci$once.per.year}
}
\value{
a list of \code{double[nVXaSESRISKHIVTB x nAGES]} matrices of derivatives of state variables due to incidence i.e. dY (net), dY.in (inflow) and dY.out (outflow)
}
\description{
Calculates the derivative dY/dt of matrix Y due to incidence split into negative and positive contributions (outflow and inflow)
}
