% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{derivs.inci}
\alias{derivs.inci}
\title{Calculates the derivative dY/dt of matrix Y due to incidence}
\usage{
derivs.inci(t = NA, Y = NA, p = NA, p.inci = NA, once = F)
}
\arguments{
\item{t}{current time}

\item{Y}{a double[nVXaSESRISKHIVTB x nAGES] matrix of state variables}

\item{p}{an environment with initialized model parameters}

\item{p.inci}{incidence data structure including incidence matrices etc}

\item{once}{logical that should match p.inci$once.per.year}
}
\value{
a double[nVXaSESRISKHIVTB x nAGES] matrix of derivatives of state variables due to incidence
}
\description{
Calculates the derivative dY/dt of matrix Y due to incidence
}
