% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{derivs.Tm.in.out}
\alias{derivs.Tm.in.out}
\title{Calculates the derivative dY/dt of matrix Y due to Mtb transmission split into negative and positive derivativs (outflow and inflow)}
\usage{
derivs.Tm.in.out(t, Y, parms)
}
\arguments{
\item{t}{current time}

\item{Y}{a double[nVXaSESRISKHIVTB x nAGES] matrix of state variables}

\item{parms}{an environment with initialized model parameters}
}
\value{
a list of dY (net flow), dY.in (inflow), dY.out (outflow)
}
\description{
\code{derivs.Tm()} calculates the contribution of Mtb transmission to the derivatives of \code{Y} at time \code{t} with parameters \code{parms} 
Within \code{derivs.Tm()} the function \code{foi()} is called to calculate the force of infection.
}
