% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{derivs.Tm}
\alias{derivs.Tm}
\title{Calculates the derivative dY/dt of matrix Y due to Mtb transmission}
\usage{
derivs.Tm(t, Y, parms)
}
\arguments{
\item{t}{current time}

\item{Y}{a double[nVXaSESRISKHIVTB x nAGES] matrix of state variables}

\item{parms}{an environment with initialized model parameters}
}
\value{
a \code{double [nVXaSESRISKHIVTB x nAGES]} of derivatives of state variables
}
\description{
\code{derivs.Tm()} calculates the contribution of Mtb transmission to the derivatives of \code{Y} at time \code{t} with parameters \code{parms} 
Within \code{derivs.Tm()} the function \code{foi()} is called to calculate the force of infection.
}
