% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-query-fncs-v5.R
\name{eval.output.vs.targets}
\alias{eval.output.vs.targets}
\title{calculates target model values from the output generated by running the model}
\usage{
eval.output.vs.targets(output, targets, p = NULL)
}
\arguments{
\item{output}{the output from running the model: a list with elements stocks and flows}

\item{targets}{a data frame with targets (from reading the targets csv)}

\item{p}{initialized model parameters}
}
\value{
the targets data frame with new columns with values calculated from the model output
}
\description{
this function calls calc.targets() to do calculate non-combined targets and then calculates combined targets from the calc.targets() output
}
