% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-age-related-fncs-v14.R
\name{expand.contact.matrix}
\alias{expand.contact.matrix}
\title{Reformats the contacts matrix with 5 year age groups A0..A75 (as read from Polymod data) into the appropriate format}
\usage{
expand.contact.matrix(parms, contactM)
}
\arguments{
\item{parms}{the initialized parameters environment}

\item{contactM}{contacts matrix with 5 year age groups}
}
\value{
the contacts matrix in the appropriate format i.e. parms$nAGES x parms$nAGES
}
\description{
Reformats the contacts matrix with 5 year age groups A0..A75 (as read from Polymod data) into the appropriate format
}
