% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{fitted.parameters}
\alias{fitted.parameters}
\title{Selects non-constant parameters (to be fitted) from a data frame}
\usage{
\method{fitted}{parameters}(parameters.df = NULL)
}
\arguments{
\item{parameters.df}{a data frame with parameters}
}
\value{
a data frame with non-constant parameters (to be fitted)
}
\description{
Selects non-constant parameters (to be fitted) from a data frame
}
