% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{foi}
\alias{foi}
\title{Calculate the force of infection for Mtb transmission}
\usage{
foi(t, Y, parms)
}
\arguments{
\item{t}{current time}

\item{Y}{a \code{double[nVXaSESRISKHIVTB x nAGES]} matrix of state variables}

\item{parms}{an environment with initialized model parameters}
}
\value{
a \code{double [nAGES]} of the force of infection by age group
}
\description{
\code{foi()} calculates the force of infection for Mtb transmission taking into account the age dependent contact matrices
}
