% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{fractions.by.stage.at.birth}
\alias{fractions.by.stage.at.birth}
\title{Sets the fractions of newborns (by state)}
\usage{
fractions.by.stage.at.birth(p)
}
\arguments{
\item{p}{params environment}
}
\value{
a vector of fractions by state
}
\description{
Sets the fractions of newborns in each state using the XML fraction.at.birth attribute of 
the VXa, SES, RISK, HIV and TB dimensions
This distribution is used to generate an appropiate distribution of newborns over states
}
