% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-fncs-v19.R
\name{generate.all.flows.output}
\alias{generate.all.flows.output}
\title{generates the output for all flows}
\usage{
generate.all.flows.output(inci, params)
}
\arguments{
\item{inci}{incidence output (result from incidence.from.model.run())}

\item{params}{initialized model parameters}
}
\value{
a data frame with flows output
}
\description{
generates the output for all flows
}
