% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-fncs-v19.R
\name{generate.inci.output.df}
\alias{generate.inci.output.df}
\title{a helper function called by generate.single.dim.flows.output() to generate a data frame with incidence output for a specific dimension}
\usage{
generate.inci.output.df(t, inci.dim, dim, subjectname, matname, params, cols)
}
\arguments{
\item{t}{vector of output time points}

\item{inci.dim}{list of matrices of various types of incidence (TBp, Tm etc) split out into dY, dY.in, dY.out for all time points}

\item{dim}{the dimension to generate incidence output for}

\item{subjectname}{subject e.g. "Tm" "TBp" etc}

\item{matname}{matrix name i.e. "dY.in" or "dY.out"}

\item{params}{initialized model parameters}

\item{cols}{set of cols of dimensions with state names by dimension}
}
\value{
a data table with incidence output for a specific subject by year and age group
}
\description{
a helper function called by generate.single.dim.flows.output() to generate a data frame with incidence output for a specific dimension
}
