% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-fncs-v19.R
\name{generate.prev.output.df}
\alias{generate.prev.output.df}
\title{helper function called by generate.prevalence.output() to generate prevalence output as a data frame}
\usage{
generate.prev.output.df(t, state, params, cols)
}
\arguments{
\item{t}{vector of output time points}

\item{state}{list of matrices of state variables x age groups by time points}

\item{params}{initialized model parameters}

\item{cols}{names of states by individual dimension}
}
\value{
a data frame with prevalence output
}
\description{
helper function called by generate.prevalence.output() to generate prevalence output as a data frame
}
