% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-fncs-v19.R
\name{generate.prevalence.output}
\alias{generate.prevalence.output}
\title{generates prevalence output}
\usage{
generate.prevalence.output(t, state, params)
}
\arguments{
\item{t}{vector of output time points}

\item{state}{list of matrices of state variables x age groups by time points}

\item{params}{initialized model parameters}
}
\value{
a data table with prevalence output by year and age group
}
\description{
generates prevalence output
}
