% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-fncs-v19.R
\name{generate.single.dim.flows.output}
\alias{generate.single.dim.flows.output}
\title{generates the flows output for a single dimension}
\usage{
generate.single.dim.flows.output(t, inci.dim, params, dim = NA)
}
\arguments{
\item{params}{initialized model parameters}

\item{dim}{dimension name e.g. "TB"}

\item{dim.inci}{incidence output for a single dimension (result from incidence.from.model.run())}
}
\value{
a data frame with flows output
}
\description{
generates the flows output for a single dimension
}
