% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-age-related-fncs-v14.R
\name{get.age.indices}
\alias{get.age.indices}
\title{Creates a list of vectors of indices of age groups (taking into account age groups size)}
\usage{
get.age.indices(p, age.groups)
}
\arguments{
\item{p}{parameters environment}

\item{age.groups}{lower limits of age groups e.g. a named vector with names c(A0, A15) and values c(0, 15)}
}
\value{
aging matrix
}
\description{
Creates a list of vectors of indices of age groups (taking into account age groups size)
}
