% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{get.bgdeathrate}
\alias{get.bgdeathrate}
\title{Returns the background deathrate for time t based on a baseline run (see new.demography.from.model.run())}
\usage{
get.bgdeathrate(p, t)
}
\arguments{
\item{p}{params environment}

\item{t}{time}
}
\value{
a vector of background deathrate by age at time t
}
\description{
The background deathrate is the difference between the WHO deaths and the TB HIV deaths divided by population size.
See also the vignette on output.
}
