% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{get.deathrate.allcauses}
\alias{get.deathrate.allcauses}
\title{Returns the deathrate for time t (read from WHO data using deathrate.from.data())}
\usage{
get.deathrate.allcauses(p, t)
}
\arguments{
\item{p}{params environment}

\item{t}{time}
}
\value{
a vector of deathrates by age group at time t
}
\description{
Returns the deathrate for time t (read from WHO data using deathrate.from.data())
}
