% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-data-reading-fncs-v13.R
\name{get.demography}
\alias{get.demography}
\title{Reads a csv file with demographics data (from UNPOP) for a country by year (in rows) and single age years (in columns)}
\usage{
get.demography(p, fname = NA, countrycode = NA)
}
\arguments{
\item{p}{an environment with parameters that is used to convert the population data to the age groups in the parameters}

\item{fname}{the file name}

\item{countrycode}{the 3 letter ISO country code}
}
\value{
a matrix with the population by year (rownames) and age group (colnames)
}
\description{
Reads a csv file with demographics data (from UNPOP) for a country by year (in rows) and single age years (in columns)
}
