% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{get.seeded.infections}
\alias{get.seeded.infections}
\title{reads the fractions of seeded TB infections by stage from <TB><seeded.infections>}
\usage{
get.seeded.infections(xmldoc)
}
\arguments{
\item{xmldoc}{the XML document}
}
\value{
a data frame with seeded infections ; columns stage, age.from, age.thru, fraction
}
\description{
reads the fractions of seeded TB infections by stage from <TB><seeded.infections>
}
