% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{get.target.hits}
\alias{get.target.hits}
\title{Reads a targets file, evaluates the output versus the targets and adds columns target hits to the targets file and returns a data frame}
\usage{
get.target.hits(output = NULL, model.params = NULL)
}
\arguments{
\item{output}{model output}

\item{model.params}{environment with initialized model parameters}
}
\value{
the model output with target hits added as a list element
}
\description{
Reads a targets file, evaluates the output versus the targets and adds columns target hits to the targets file and returns a data frame
}
