% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{inci.data.frame.from.xpath}
\alias{inci.data.frame.from.xpath}
\title{parses the incidence data specification for a dimension}
\usage{
inci.data.frame.from.xpath(xml, xpath)
}
\arguments{
\item{xml}{the XML document}

\item{xpath}{the path to the incidence specification in the XML e.g. <VXa><VXa.incidence>}
}
\value{
a data frame with specification of incidence data ; columns file, times, values, proportions, denominator, once.per.year
}
\description{
parses the incidence data specification for a dimension
}
