% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-fncs-v19.R
\name{incidence.from.model.run}
\alias{incidence.from.model.run}
\title{Regenerates incidence output from model output}
\usage{
incidence.from.model.run(out, params)
}
\arguments{
\item{out}{model output from run.deSolve() i.e. a list of matrices (state variables by age group) as a function of time}

\item{params}{environment with initialized model parameters}
}
\value{
a list with elements t, PREV, and depending on output options e.g. TB, HIV, VXa with 
t:    a vector of time points 
PREV: a list of matrices by time point with prevalence
TB:   a list of incidence elements by time point with TB progression (TBp), TB transmission(Tm) and the various TB treatment matrices (e.g. TBtr_init)
      each incidence element contains dY, dY.in and dY.out
HIV, VXa, etc: cf. TB
}
\description{
Regenerates incidence output from model output
}
