% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-fncs-v19.R
\name{incidence.from.Y.matrix}
\alias{incidence.from.Y.matrix}
\title{Helper function called by incidence.from.model.run()}
\usage{
incidence.from.Y.matrix(t, Y, p, sel.output = NA)
}
\arguments{
\item{Y}{a matrix with state variables}

\item{p}{environment with initialized model parameters}
}
\value{
a list with incidence elements such as Tm, TBp, TBdeaths, TBtr_init, etc each with dY, dY.in and dY.out
}
\description{
Helper function called by incidence.from.model.run()
}
