% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-data-reading-fncs-v13.R
\name{init.incidence}
\alias{init.incidence}
\title{Initializes an environment from an incidence file and options specified}
\usage{
init.incidence(
  p,
  fname = NA,
  countrycode = NA,
  times = NULL,
  values = NULL,
  proportions = F,
  denominator = NA,
  once.per.year = F,
  target.prevalence = F
)
}
\arguments{
\item{p}{an environment with initialized parameters}

\item{fname}{the name of the tab delimited .txt file with incidence}

\item{countrycode}{3 letter ISO country code}

\item{times}{a series of times}

\item{values}{a series of values or an expression to be converted to a function producing values as a function of time}

\item{proportions}{a logical to indicate if the numbers in the file are proportions or rates (the default)}

\item{denominator}{the denominator to be used in incidence calculations either 'susc' or 'all'}

\item{once.per.year}{a logical to indicate if incidence numbers are rates (F) or fractions (T) that are moved to the new state once per year}
}
\value{
an environment that is used in the derivs.inci.common() function in TBVx-derivs-v##.R with the inci.matrix as most important part
}
\description{
The times and values arguments are converted to a function in the TBVx-initialization-funcs-v##.R
}
