% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{init.parameters.from.xml}
\alias{init.parameters.from.xml}
\title{initializes an environment with parameters for infectivity, transmission (Tm), progression, treatment}
\usage{
init.parameters.from.xml(
  p,
  xmlpath,
  xmlparameter,
  xmltransitions,
  xmlauxparameters,
  xmltimedparams = NA,
  values = F
)
}
\arguments{
\item{xmlpath}{the path to the XML parameters e.g. "//TB/TB.transmission"}

\item{xmlparameter}{the name of the XML parameter elements e.g. "TB.parameter"}

\item{xmltransitions}{a node list of XML transitions e.g. from "//TB/TB.transmission/transition.matrix/transition"}

\item{xmlauxparameters}{a node list of XML parameters defined in the transition matrix (if any) i.e. "//TB/TB.transmission/transition.matrix/parameter"}

\item{values}{a logical to indicate if values should be read for stages (e.g. for TB infectivity) instead of for transitions (from, to)}

\item{xmltimedparameters}{a node list of XML timed parameters i.e. time dependent multipliers from e.g. "//TB/TB.transmission/transition.matrix/multiplier"}
}
\value{
an environment with xmlpath, xmlparameter, dim.name, col.names, age.groups, age.ranges, parameter.matrices, timed.paramters
}
\description{
initializes an environment with parameters for infectivity, transmission (Tm), progression, treatment
}
