% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{initialize.aggregated.TBHIVptr}
\alias{initialize.aggregated.TBHIVptr}
\title{Combines TB progression, TB treatment, HIV progression, HIV treatment 
if matrices are constant over time and if age groups are the same.
Flags such as HIVp$aggregated and for each treatment matrix HIVtr[[i]]$aggregated 
keep track of aggregation status.}
\usage{
initialize.aggregated.TBHIVptr(
  TBp = NULL,
  TBtr = NULL,
  HIVp = NULL,
  HIVtr = NULL
)
}
\arguments{
\item{TBp}{TB progression matrices}

\item{TBtr}{TB treatment matrices}

\item{HIVp}{HIV progression matrices}

\item{HIVtr}{HIV treatment matrices}
}
\value{
aggregated TB and HIV progression and treatment matrices
}
\description{
Combines TB progression, TB treatment, HIV progression, HIV treatment 
if matrices are constant over time and if age groups are the same.
Flags such as HIVp$aggregated and for each treatment matrix HIVtr[[i]]$aggregated 
keep track of aggregation status.
}
