% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{initialize.incidence}
\alias{initialize.incidence}
\title{Initializes incidence data structures (TB, HIV and VXa) from XML and data files referenced in XML}
\usage{
initialize.incidence(paths = NULL, p = NULL)
}
\arguments{
\item{paths}{paths to various files ; see set.paths() function}

\item{p}{params environment to be updated}
}
\value{
environment of model parameters with initialized incidence lists
}
\description{
See also the init.incidence() function in TBVx-data-reading-fncs-v##.R
}
