% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{initialize.model.parameters}
\alias{initialize.model.parameters}
\title{Initialization of the model parameters}
\usage{
initialize.model.parameters(params = NULL)
}
\arguments{
\item{params}{An environment with the result of calling the function read.model.parameters(); 
the params argument contains e.g. data frames from files read but needs to be parsed further}
}
\value{
an environment with initialized data structures required to run the simulation 

\strong{Description of the environment returned:}
\describe{
\item{\code{AGES: double [16] 0 5 10 15 ...}}{with lower limits of 5 year age groups}
\item{\code{aging.matrix: double [16 x 16] -0.2 0.2 0 0 0 0 0 0 0 0 ...   }}{a matrix used to move the population to the next higher age group once a year ; i.e. the result of multiplying the 16x16 aging matrix with the (in this case) 240 x 16 matrix of state variables}
\item{\code{ALIVE: logical [240] TRUE TRUE TRUE TRUE TRUE TRUE ...}}{logical vector that defines the ‘alive’ states, derived from state names not ending in ‘dead'}
\item{\code{at.birth: double [240] 0.4 0 0 0 0 0 0 0 0 0 ...}}{a vector with the distribution of newborns over states (sums to 1.0)}
\item{\code{ATBIRTH: list[5]}}{distributions of newborns by state used to calculate the \code{at.birth} vector ; for instance:}
\itemize{
\item \code{VXa : double [1] 1}
\item \code{SES : double [2] 0.4 0.6}
\item \code{RISK: double [1] 1}
\item \code{HIV : double [10] 1 0 0 0 0 0 0 0 0 0}
\item \code{TB  : double [12] 1 0 0 0 0 0 0 0 0 0 0 0}
}
\item{\code{at.start: double [240 x 16] 0.321475 0.037363 0.002855 0.030072 0.000481 ... ...}}{distribution of the population by state for each of the age groups at the start of the simulation ; each age group (column) sums to 1.0}
\item{\code{birthrate: function (v)}}{a function that calculates the row index from the current year}
\item{\code{contacts: list[3]}}{}
\itemize{
\item \code{defined: logical [1] TRUE}
\item \code{M      :Formal class 'dgeMatrix' [package "Matrix"] with 4 slots}
\item \code{MplustM:Formal class 'dgeMatrix' [package "Matrix"] with 4 slots}
\item M is the contacts matrix (from the paper by Kiesha Prem et al) M is used to rebalance the contacts matrix to the updated population composition which is done once per year. See the function update.contact.matrix(parms, alive.pop) in TBVx-age-related-fncs-v##.R)
}
\item{\code{DAYSPERYEAR: double [1] 365.2425}}{number of days per year}
\item{\code{deathfn: function (v)}}{function to select the deathrate vector of the current year from the deathrate matrix (below)}
\item{\code{deathrate: double [151 x 16] 0.0527 0.0527 0.0527 0.0527 0.0512 ...}}{deathrate by year (row) and age group (column) derived from data (e.g. IND_deathrates.csv)}
\item{\code{DIMLENGTHS: integer [5] 1 2 1 10 12}}{lengths of the dimensions}
\item{\code{DIMNAMES: character [5] "VXa" "SES" "RISK" "HIV" "TB"}}{names of the dimensions}
\item{\code{DIMNAMESLIST: list[5]}}{list of state names by dimension}
\itemize{
\item \code{VXa : character [1] "never"}
\item \code{SES : character [2] "low" "high"}
\item \code{RISK: character [1] "risk1"}
\item \code{HIV : character [10] "HIV-" "HIVu1" "HIVu2" "HIVd1" ...}
\item \code{TB  : character [12] "Un" "Uc" "Lf" "Ls" ...}
}
\item{\code{hash: character [1] e.g. "9f1dfb8bec73a9aaa1a92064f9a924a9"}}{hash of this environment of initialized parameters}
\item{\code{HIV: character [10] "HIV-" "HIVu1" "HIVu2" "HIVd1" "HIVd2" "ARTn1" "ARTn2" "ARTs1" "ARTs2" ...}}{names of HIV states}
\item{\code{HIVp: environment [10]}}{environment with HIV progression parameters}
\item{\code{HIVTB: character [120] "HIV--Un" "HIV--Uc" "HIV--Lf" "HIV--Ls" "HIV--Ds" "HIV--Dc" "HIV--T" ...}}{names of the combination of HIV and TB states}
\item{\code{HIVtr: list[4]}}{list of HIV treatment environments; the names of these environments are derived from the XML}
\itemize{
\item \code{diag       : environment}
\item \code{treat      : environment}
\item \code{supp       : environment}
\item \code{HIVARTdead : environment}
}
\item{\code{inci: list[1]}}{list of incidence lists ; one for each supported dimension. Used for generating incidence matrices (function create.incidence.matrix()). Each of the list elements is an environment with all that is required to properly execute the transitions specified in the incidence files}
\item{\code{inci.dim.names: list[5]}}{list of dimensions that (currently) support incidence files}
\itemize{
\item \code{VXa: character [240] "never" "never" "never" "never" ...}
\item \code{[[2]]   : NULL}
\item \code{[[3]]   : NULL}
\item \code{HIV: character [240] "HIV-" "HIV-" "HIV-" "HIV-" ...}
\item \code{[[5]]   : NULL}
}
\item{\code{Infc: environment [11]}}{ Environment with parameters related to TB infectiousness ; most important is the matrix with infectiousness by state (a 240x240 matrix in this case)}
\item{\code{intervention: logical [1] FALSE}}{logical to flag intervention run}
\item{\code{intervention.start: double [1] 1e+06}}{default value of intervention start ; overwritten if applicable}
\item{\code{intervention.start.from: logical [1] NA}}{logical indicating if the intervention start is derived from incidence file}
\item{\code{iHIV: double [10] 1 0 0 0 0 0 0 0 0 0}}{distribution of newborns over HIV states}
\item{\code{iRISK :  double [1] 1}}{distribution of newborns over RISK states}
\item{\code{iSES :  double [2] 0.4 0.6}}{distribution of newborns over SES states}
\item{\code{iTB :  double [12] 1 0 0 0 0 0 0 0 0 0 ...}}{distribution of newborns over TB states}
\item{\code{iVXa :  double [1] 1}}{distribution of newborns over VXa states}
\item{\code{iHIV, iRISK, iSES, iTB, and iVXa }}{are used to generate ATBIRTH}
\item{\code{nAGES :  integer[1] 16}}{}
\item{\code{nHIV :  integer[1] 10}}{}
\item{\code{nHIVTB :  integer[1] 120}}{}
\item{\code{nRISK :  integer[1] 1}}{}
\item{\code{nRISKHIVTB :  integer[1] 120}}{}
\item{\code{nSES :  integer[1] 2}}{}
\item{\code{nSESRISKHIVTB :  integer[1] 240}}{}
\item{\code{nTB :  integer[1] 12}}{}
\item{\code{nVXa :  integer[1] 1}}{}
\item{\code{nVXaSESRISKHIVTB :  integer[1] 240}}{}
\item{}{nAGES, nHIV, nHIVTB, nRISK, nRISKHIVTB etceteta are the sizes of (combined) dimensions}
\item{\code{paths : environment [13]}}{environment with paths to various files ; see set.paths() function}
\item{\code{popfname :  character[1] e.g. "./countries/ZAF/data/demographics.csv"}}{filename of UNPOP demographics file (i.e. population development over time)}
\item{\code{PROGRESSION : list[5]}}{list with HIV, RISK, SES and VXa progression and in addition HIV treatment. Produced by PROGRESSION = list(HIV=HIVp,HIVtr=HIVtr,RISK=RISKp,SES=SESp,VXa=VXap)  where HIVp etc have been initialized separately. TB progression and treatment are in TBp and TBtr}
\itemize{
\item \code{HIV : environment}
\item \code{HIVtr: list[4]}
\item \code{RISK : NULL}
\item \code{SES  : NULL}
\item \code{VXa  : NULL}
}
\item{\code{RISK : character "risk1"}}{names of RISK states}
\item{\code{RISKHIVTB : character [120] "risk1-HIV--Un" "risk1-HIV--Uc" "risk1-HIV--Lf" "risk1-HIV--Ls" ...}}{names of combined RISKHIVTB states}
\item{\code{RISKp :  NULL}}{RISK progression}
\item{\code{run.params : environment}}{Environment with parameters that are related to running the simulation rather than the model itself (i.e. states, transitions etc.)}
\item{\code{SES :  character [2] "low" "high"}}{names of RISK states}
\item{\code{SESp :  NULL}}{SES progression}
\item{\code{SESRISKHIVTB :  character [240] "low-risk1-HIV--Un" "low-risk1-HIV--Uc" "low-risk1-HIV--Lf" ...}}{names of combined SESRISKHIVTB states}
\item{\code{TB :  character [12] "Un" "Uc" "Lf" "Ls" "Ds" "Dc" "T" "R" "TBdead" "TBHIVdead" "Rdead" ...}}{names of TB states}
\item{\code{TBHIVptr : environment[3]}}{TBHIV progression NOTE: to be explained!!}
\itemize{
\item \code{age.groups: integer[2]}
\item \code{age.ranges: list[2]}
\item \code{parameter.matrices : list[2]}
}
\item{\code{TBp : environment[10]}}{TB progression}
\item{\code{TBtr : list[5]}}{list of TB treatment matrices}
\itemize{
\item \code{RDs    : environment[10]} 
\item \code{LfLsDs : environment[10]}
\item \code{init   : environment[10]}
\item \code{nodead : environment[10]}
\item \code{died   : environment[10]}
}
\item{\code{Tm : environment[9]}}{TB transmission}
\item{\code{unpopdem :  double [151 x 16] 2024 2106 2218 2326 2419 ...}}{UNPOP population by year and age group (i.e. adjusted to modelled age groups)}
\item{\code{unpopdemfn : function (v)}}{function that returns the row index for the current year}
\item{\code{VXa :  character[1] "never"}}{names of VXa states}
\item{\code{VXap :  NULL}}{VXa progression}
\item{\code{VXaSESRISKHIVTB :  character [240] "never-low-risk1-HIV--Un" "never-low-risk1-HIV--Uc" ...}}{names of combined VXaSESRISKHIVTB states}
\item{\code{xml : list[2]}}{list with XML document and XML Schema}
\itemize{
\item \code{doc   :list[2]}
\item \code{schema:list[2]}
}
\item{\code{xmlfilename :  character[1] "./countries-examples/ZAF/parameters/XMLinput_2910_m.xml"}}{file name of XML input file}
\item{\code{y.ini :  double [240 x 16] 650.584 75.613 5.779 60.859 0.974 ...}}{initial population by state and age group}
\item{\code{y.ini.read :  double [240 x 16] 2024 0 0 0 0 ...}}{the initial population read from \code{demographics.csv} for the start year of the simulation, before seeding ; seeding creates \code{y.ini}}
}
}
\description{
initializes the model parameters
}
