% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{initialize.progression}
\alias{initialize.progression}
\title{Initialize progression}
\usage{
initialize.progression(p = NULL, dim = NA)
}
\arguments{
\item{p}{params environment}

\item{dim}{(a string to specify the dimension to initialize progression for}
}
\value{
an environment to be assigned to TBp, HIVp etc
}
\description{
initializes VXa, SES, RISK, HIV, or TB progression from the XML specification
}
