% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{initialize.TB.infectivity}
\alias{initialize.TB.infectivity}
\title{initializes TB infectivity from the XML specification}
\usage{
initialize.TB.infectivity(p = NULL)
}
\arguments{
\item{p}{params environment}
}
\value{
an environment to be assigned to Infc
}
\description{
initializes TB infectivity from the XML specification
}
