% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{initialize.TB.transmission}
\alias{initialize.TB.transmission}
\title{initializes TB transmission from the XML specification}
\usage{
initialize.TB.transmission(p = NULL)
}
\arguments{
\item{p}{params environment}
}
\value{
an environment to be assigned to Tm
}
\description{
initializes TB transmission from the XML specification
}
