% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{initialize.treatment}
\alias{initialize.treatment}
\title{initializes HIV or TB treatment from the XML specification}
\usage{
initialize.treatment(p = NULL, dim = NA)
}
\arguments{
\item{p}{params environment}

\item{dim}{a string to specify the dimension to initialize treatment for}
}
\value{
an environment to be assigned to TBtr, HIVtr etc
}
\description{
initializes HIV or TB treatment from the XML specification
}
