% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-age-related-fncs-v14.R
\name{matmul.by.age.group.in.out}
\alias{matmul.by.age.group.in.out}
\title{Multiples the age specific matrices in subject with the appropriate subset of Y to produce one matrix of the same dimensions as Y 
and splits the result in net, positive and negative contributions dependent on the sign of the multiplying age group matrix}
\usage{
matmul.by.age.group.in.out(subject, Y, mult = 1)
}
\arguments{
\item{subject}{an environment from which the parameter matrices are derived}

\item{Y}{a matrix with states in rows and ages in columns}

\item{mult}{a multiplier to be applied to the resulting matrix}
}
\value{
a list of dY (net), dY.in (positive contributions) and dY.out (negative contributions)
}
\description{
Multiples the age specific matrices in subject with the appropriate subset of Y to produce one matrix of the same dimensions as Y 
and splits the result in net, positive and negative contributions dependent on the sign of the multiplying age group matrix
}
