% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-age-related-fncs-v14.R
\name{matmul.by.age.group}
\alias{matmul.by.age.group}
\title{Multiples the age specific matrices in subject with the appropriate subset of Y to produce one matrix of the same dimensions as Y}
\usage{
matmul.by.age.group(subject, Y, mult = 1)
}
\arguments{
\item{subject}{an environment from which the parameter matrices are derived}

\item{Y}{a matrix with states in rows and ages in columns}

\item{mult}{a multiplier to be applied to the resulting matrix}
}
\value{
the result of the multiplication of subsets of Y with the appropriate matrix for an age group
}
\description{
Multiples the age specific matrices in subject with the appropriate subset of Y to produce one matrix of the same dimensions as Y
}
