% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-age-related-fncs-v14.R
\name{matmul.for.one.age.group}
\alias{matmul.for.one.age.group}
\title{Matrix multiplication of X and Y}
\usage{
matmul.for.one.age.group(X, Y)
}
\arguments{
\item{X}{a square parameters matrix for one age group}

\item{Y}{a matrix (of state by ages)}
}
\value{
the matrix result of X and Y
}
\description{
Matrix multiplication of X and Y
}
