% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-data-reading-fncs-v13.R
\name{matrix.from.popdf}
\alias{matrix.from.popdf}
\title{Reads a tab delimited file with population composition as produced by a previous simulation run (e.g. as in output$population)}
\usage{
matrix.from.popdf(p, pop = NULL)
}
\arguments{
\item{fname}{the name of a tab delimited file with columns year, country code, and age group (0 thru 79, 80, 90)}
}
\value{
a matrix with population composition by year (in rownames) and age groups in columns
}
\description{
Checks that age groups are the same as in the parameters environment
}
