% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{merge.stocks.and.flows}
\alias{merge.stocks.and.flows}
\title{Merges stocks and flows output}
\usage{
\method{merge}{stocks.and.flows}(output)
}
\arguments{
\item{output}{model output}
}
\value{
data frame with merged stocks and flows model output
}
\description{
Merges stocks and flows output
}
