% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{modelversion}
\alias{modelversion}
\title{Returns the model version as a string}
\usage{
modelversion()
}
\value{
a string with the model version e.g. "3_3_3"
}
\description{
Returns the model version as a string
}
