% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{modify.input.csv}
\alias{modify.input.csv}
\title{Modifies the csv file with parameters that will replace parameters in the XML}
\usage{
modify.input.csv(input.csv = NULL, new.parameter.values = NULL)
}
\arguments{
\item{input.csv}{a data frame read from the file indicated by paths$parameters}

\item{new.parameter.values}{a named vector of parameter values to replace the non constant parameters in input.csv}
}
\value{
the modified input.csv
}
\description{
Modifies the csv file with parameters that will replace parameters in the XML
}
