% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{new.demography.from.model.run}
\alias{new.demography.from.model.run}
\title{Creates demography output from a model run}
\usage{
new.demography.from.model.run(out, params)
}
\arguments{
\item{out}{}

\item{params}{an environment with initialized model parameters}
}
\value{
a \code{list} with \code{pop} i.e. population composition, \code{dHIVTBx} i.e. number of HIV and TB deaths per year, 
        \code{dBGx} number of background deaths per year, \code{dPOPadj} fractional population adjustment rates (fraction of the population per year)
}
\description{
Creates demography output from a model run
}
