% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{optimize.params}
\alias{optimize.params}
\title{Attempts to combine age dependent parameter matrices to a single parameter matrix}
\usage{
optimize.params(Ti)
}
\arguments{
\item{Ti}{an environment with a list of parameter matrices by age grouping}
}
\value{
updated environment with updated list of parameter matrices
}
\description{
Attempts to combine age dependent parameter matrices to a single parameter matrix
}
