% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{output.options.from.xpath}
\alias{output.options.from.xpath}
\title{a helper function to parse the <output><detailed.output><flows> elements}
\usage{
output.options.from.xpath(xmldoc, xpath)
}
\arguments{
\item{xmldoc}{the XML document}

\item{xpath}{by default "//output/detailed.output/flows"}
}
\value{
an data frame with output options by dimension
}
\description{
a helper function to parse the <output><detailed.output><flows> elements
}
