% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{parse.fixed.parameters.from.xml}
\alias{parse.fixed.parameters.from.xml}
\title{parse fixed parameter values from XML into a data frame}
\usage{
parse.fixed.parameters.from.xml(
  p,
  xmlpath,
  xmlparameter,
  col.names = NULL,
  age.groups = NULL,
  dim.name
)
}
\arguments{
\item{p}{initialized parameters environment}

\item{xmlpath}{path to the XML elements e.g. "//TB/progression"}

\item{xmlparameter}{name of the parameter e.g. "TB.parameter"}

\item{col.names}{column names to use for parameter value dependencies}

\item{age.groups}{a named vector of age group lower bounds}

\item{dim.name}{the dimension e.g. "TB"}
}
\value{
a data frame with parameter names and values including dependencies ; column names as passed in col.names
}
\description{
parse fixed parameter values from XML into a data frame
}
