% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{parse.output.spec}
\alias{parse.output.spec}
\title{parses the contents of the XML element <output><detailed.output>}
\usage{
parse.output.spec(xmldoc, xmlpath = "//output/detailed.output")
}
\arguments{
\item{xmldoc}{the XML document}

\item{xmlpath}{by default "//output/detailed.output"}
}
\value{
an environment with details on which output to generate and when ; see XML
}
\description{
parses the contents of the XML element <output><detailed.output>
}
