% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{parse.time.series}
\alias{parse.time.series}
\title{parses a series of times and values derived from the incidence list elements inci.values and inci.times}
\usage{
parse.time.series(p, inci, fixed.params)
}
\arguments{
\item{p}{the parameters environment}

\item{inci}{a list of incidence parameters}

\item{fixed.params}{fixed parameters used in parsing the 'values' expression}
}
\value{
a function of time for multiplying the incidence with
}
\description{
parses a series of times and values derived from the incidence list elements inci.values and inci.times
}
