% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{parse.timed.parameters}
\alias{parse.timed.parameters}
\title{parses XML timed parameters such as <multiplier name="etamul" times="1960:2020" values="1/(1+exp(-sh*(x-midx)))"/> <multiplier>}
\usage{
parse.timed.parameters(p, xmlnodes, fixed.params)
}
\arguments{
\item{p}{the parameters environment}

\item{xmlnodes}{a node list of XML parameters with name, times and values attributes}

\item{fixed.params}{fixed parameters used in parsing the 'values' expression}
}
\value{
an environment with named functions of time
}
\description{
parses XML timed parameters such as <multiplier name="etamul" times="1960:2020" values="1/(1+exp(-sh*(x-midx)))"/> <multiplier>
}
