% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{query.from.col.names}
\alias{query.from.col.names}
\title{generates a query based on the column names passed and the index within each of the dimensions (VXa, SES, etc)}
\usage{
query.from.col.names(
  parms = NULL,
  col.names = NULL,
  vxa,
  ses,
  risk,
  hiv,
  tb,
  ag
)
}
\arguments{
\item{parms}{the initialized params environment}

\item{col.names}{a vector of column names (e.g. "SES.stage","HIV.stage", etc) to generate the query from}

\item{vxa}{index of the VXa stage to generate the query for}

\item{ses}{index of the SES stage to generate the query for}

\item{risk}{index of the RISK stage to generate the query for}

\item{hiv}{index of the HIV stage to generate the query for}

\item{tb}{index of the TB stage to generate the query for}

\item{ag}{index of the age group to generate the query for}
}
\value{
a query to be used in generating a small parameters matrix
}
\description{
called by create.small.matrix.from.query()
}
