% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-query-fncs-v5.R
\name{query.from.target}
\alias{query.from.target}
\title{generates a query for a target ; either the numerator or the denominator}
\usage{
query.from.target(target = NA, p = NULL)
}
\arguments{
\item{target}{one target (i.e. either numerator or denominator in the targets file)}

\item{p}{initialized model parameters}
}
\value{
a string with the generated query
}
\description{
generates a query for a target ; either the numerator or the denominator
}
