% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-data-reading-fncs-v13.R
\name{read.contact.matrix}
\alias{read.contact.matrix}
\title{Reads a tab delimited file with a contacts matrix}
\usage{
read.contact.matrix(fname)
}
\arguments{
\item{fname}{the name of a tab delimited file with 16 rows and 16 columns with names A0, A5, A10 ... A75}
}
\value{
a 16 x 16 matrix
}
\description{
Reads a tab delimited file with a contacts matrix
}
