% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{read.model.parameters}
\alias{read.model.parameters}
\title{Reads model parameters from XML}
\usage{
read.model.parameters(paths = NULL)
}
\arguments{
\item{paths}{an environment with paths to various files, country code etc. See the function set.paths()}
}
\value{
an environment with initialized constants and data structures related to demography
}
\description{
reads the model parameters (without initialization ; called by the run() function) from XML
}
