% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-data-reading-fncs-v13.R
\name{read.population}
\alias{read.population}
\title{Reads the initial population composition from a file}
\usage{
read.population(p, rownms, colnms, fname = NA, countrycode = NA, year = NA)
}
\arguments{
\item{p}{an environemnt with initialized parameters}

\item{rownms}{rownames}

\item{colnms}{colnames}

\item{fname}{the name of the .csv file}

\item{countrycode}{3 letter ISO country code}

\item{year}{the year (i.e. row in the file)}
}
\value{
a matrix with the population composition in the year specified
}
\description{
Reads the initial population composition from a file
}
