% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-query-fncs-v5.R
\name{read.targets}
\alias{read.targets}
\title{reads a csv with defined targets, checks for consistency and produces a data frame}
\usage{
read.targets(fname)
}
\arguments{
\item{fname}{file name of the csv to read}
}
\value{
a data frame with targets
}
\description{
reads a csv with defined targets, checks for consistency and produces a data frame
}
