% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-data-reading-fncs-v13.R
\name{read.xml}
\alias{read.xml}
\title{Reads and validates an XML file using the provided path to the XML Schema file}
\usage{
read.xml(xmlfile, schemapath)
}
\arguments{
\item{xmlfile}{the name of the xml file}

\item{schemapath}{the path to the XML schema (.xsd)}
}
\value{
an XML document
}
\description{
Reads and validates an XML file using the provided path to the XML Schema file
}
