% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{rk.parameters}
\alias{rk.parameters}
\title{reads the attributes from the XML element <simulation><options><numerical.integration/> including defaults from the XML Schema
to be stored in params$num.int}
\usage{
rk.parameters(xmlschema, node = NA, names = NA, types = NA)
}
\arguments{
\item{xmlschema}{the XML schema}

\item{node}{a nodelist}

\item{names}{a vector of attribute names}

\item{types}{a vector of parameter types}
}
\value{
a named list of values about the numerical integration algorithm
}
\description{
reads the attributes from the XML element <simulation><options><numerical.integration/> including defaults from the XML Schema
to be stored in params$num.int
}
