% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-matrix-fncs-v12.R
\name{roll.matrix.by.rows}
\alias{roll.matrix.by.rows}
\title{A helper function called from create.incidence.matrix()}
\usage{
roll.matrix.by.rows(M, by, negate = T)
}
\arguments{
\item{M}{a square matrix}

\item{by}{a number to roll down (by>0) or up (by<0) a copy of the matrix 'from' -1 values and (by default) negate those values}

\item{negate}{default T so the 'to' values will be +1's}
}
\value{
a double matrix of the same dimensions as M with +1s added in the appropriate positions
}
\description{
A helper function called from create.incidence.matrix()
}
