% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{run.model}
\alias{run.model}
\title{called from run()}
\usage{
run.model(model.params = NULL, output.flows = T)
}
\arguments{
\item{model.params}{an environment with initialized model parameters}

\item{output.flows}{if F no flows output is generated}
}
\value{
a \code{list} with  \code{stocks} and subsets of stocks i.e. \code{dead.999}, \code{count.999}, \code{alive.500}, \code{alive.999}
        and in addition \code{flows} (if \code{output.flows=T}) and if \code{econ.output}="true" in the XML also     
        \code{population}, \code{dHIVTBx}, \code{dBGx}, \code{dPOPadj} 
        and if vaccination active also \code{vaccinated}
}
\description{
within run.model the run.deSolve() function is called to run the model ; in addition the output of run.deSolve() is processed to produces stocks and flows output and in addition output on deaths and counts
}
