% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-optimization-fncs-v2.R
\name{run.optimizer}
\alias{run.optimizer}
\title{Runs the model and tries to find a set of parameter values that minimizes the weighted sum of squares of differences between target values and model values
Results are written to files in the logs folder}
\usage{
run.optimizer(
  paths = NULL,
  new.parameter.values = NULL,
  constant.new.parameter.values = F,
  lmcontrolcsv = NULL,
  nruns = 1,
  pwr = NULL
)
}
\arguments{
\item{new.parameter.values}{a named vector of parameter values to replace the non constant parameters in input.csv}

\item{paths=NULL}{initialized paths}
}
\description{
Runs the model and tries to find a set of parameter values that minimizes the weighted sum of squares of differences between target values and model values
Results are written to files in the logs folder
}
