% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-run-v1.R
\name{run}
\alias{run}
\title{The purpose of run() is to read the XML input file, update parameter values (if applicable), intialize the model parameters (i.e. setting up the matrices), do one model run, and return the output.}
\usage{
run(
  paths = NULL,
  new.parameter.values = NULL,
  write.to.file = F,
  write.xml = NULL,
  output.flows = T,
  combine.stocks.and.flows = F,
  baseline = NULL,
  output.format = "txt",
  sample.parameters = F
)
}
\arguments{
\item{paths=NULL}{initialized paths}

\item{new.parameter.values=NULL}{a named vector of parameter values that will overwrite (part of) the non-constant parameters in the parameters file (usually input.csv)}

\item{write.to.file=F}{a logical that determines if output will be written to files}

\item{write.xml=NULL}{the name of a file to write the modified XML to}

\item{output.flows=T}{a logical that determines flows will be calculated and included in the output}

\item{combine.stocks.and.flows=F}{if T stocks and flows will be written to a single file}

\item{baseline=NULL}{if not NULL, a reference to an output structure produced by a baseline run}

\item{output.format='txt'}{either ‘txt’ for tab-delimited text, ‘fst’ for fst, or ‘parquet’ for parquet}

\item{sample.parameters=F}{if T, values of the non-constant parameters in input,csv will be sampled from the distribution specified in the ‘distr’ column of input.csv}
}
\value{
a list with the following elements: stocks, flows, hits (only if targets is specified) and additional elements if econ.output=”true” in the XML
}
\description{
If the argument new.parameter.values is not supplied while the paths$parameters is not NA, all parameters from the parameters file (input.csv) will be used to overwrite the parameter values in the XML.
}
