% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{sample.fitted.parameters}
\alias{sample.fitted.parameters}
\title{Samples uniform values from non-constant parameters from a data frame}
\usage{
sample.fitted.parameters(selected.parameters = NULL)
}
\arguments{
\item{selected.parameters}{a data frame with parameters}
}
\value{
a named vector with sampled parameter values (non-constant parameters only)
}
\description{
Samples uniform values from non-constant parameters from a data frame
}
