% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{scale.alive.population}
\alias{scale.alive.population}
\title{Scales the current population composition and size to the initial population composition and size}
\usage{
\method{scale}{alive.population}(Y, parms)
}
\arguments{
\item{Y}{a \code{double [p$nVXaSESRISKHIVTB x p$nAGES]} of state variables of the current population}

\item{parms}{an environment with initialized model parameters}
}
\value{
a \code{double [p$nVXaSESRISKHIVTB x p$nAGES]} of state variables of the rescaled population
}
\description{
Scales the current population composition and size to the initial population composition and size
}
