% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{schemaversion}
\alias{schemaversion}
\title{Returns the XML Schema version as a string}
\usage{
schemaversion()
}
\value{
a string with XML Schema version e.g. "TB-Vx-schema-T.xsd"
}
\description{
Returns the XML Schema version as a string
}
