% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{seed.initial.population}
\alias{seed.initial.population}
\title{Seeds the initial population with fractions by state as specified in the XML}
\usage{
seed.initial.population(parms)
}
\arguments{
\item{parms}{environment with parameters}
}
\value{
an environment with parms$y.ini containing the initialized population
}
\description{
Seeds the initial population with fractions by state as specified in the XML
}
