% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{selected.parameters}
\alias{selected.parameters}
\title{Selects parameters from a data frame}
\usage{
selected.parameters(parameters.df = NULL, constant = NA)
}
\arguments{
\item{parameters.df}{a data frame with parameters}

\item{constant}{a logical to indicate if constant or non constant parameters should be selected}
}
\value{
a data frame with selected parameters
}
\description{
Selects parameters from a data frame
}
