% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-initialization-fncs-v20.R
\name{set.baseline.in.model.parameters}
\alias{set.baseline.in.model.parameters}
\title{Sets population adjustment and background death rate in intervention equal to baseline run}
\usage{
set.baseline.in.model.parameters(params = NULL, output = NULL)
}
\arguments{
\item{params}{an environment with initialized parameters of the intervention run that will be modified}

\item{output}{the output from the baseline run}
}
\description{
uses output$dPOPadj, output$popadjrateintv, output$dBGx, output$population to calculate
params$popadjrateintv, params$popadjfnintv, params$bgdeathrate and params$bgdeathfn
}
