% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-mod-xml-fncs-v2.R
\name{set.node.attrs}
\alias{set.node.attrs}
\title{assigns new values to attributes in an XML document using the xpath to locate the set of attributes}
\usage{
set.node.attrs(xmldoc, xpath, attrname = NULL, newvalues = NULL)
}
\arguments{
\item{xmldoc}{the in memory XML document}

\item{xpath}{xpath to the attributes}

\item{attrname}{attribute name}

\item{newvalues}{the new attribute values}
}
\value{
the modified XML document
}
\description{
assigns new values to attributes in an XML document using the xpath to locate the set of attributes
}
