% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{set.paths}
\alias{set.paths}
\title{set.paths() creates an R environment with file paths to the XML file, parameter file, targets file etc}
\usage{
set.paths(
  mydir = here(),
  countries = "countries",
  countrycode = NA,
  xml = NA,
  parameters = NA,
  targets = NA,
  baseline = NA,
  VXa.incidence.files = NA,
  lglevel = "DEBUG",
  taskenvvar = "TaskID"
)
}
\arguments{
\item{mydir}{the ‘root’ directory}

\item{countries}{the name of the countries folder (which will be appended to mydir)}

\item{countrycode}{3 letter ISO code of the country (e.g. ZAF)}

\item{xml}{the filename of the XML input file}

\item{parameters}{the filename of the parameters file (which may overwrite parameters in the xml inputfile)}

\item{targets}{the filename of the targets.csv file}

\item{baseline}{the filename (without extent) of the XML file of the baseline}

\item{VXa.incidence.files}{a vector of filenames (including path from the country folder) to replace the VXa incidence files specified in the XMLinput.xml file}
}
\value{
an environment with the paths to the various files and folders.
}
\description{
The only required parameters are countrycode and xml
}
\details{
This function also initializes the model log

If the parameters argument is specified, the values of the parameters designated as constant in this file (usually input.csv) will overwrite the values in the XML.

If the targets argument is specified, the output will contain a hits element with a comparison between targets and model values.
If the baseline argument is specified, the model run will use the adjustments that were made to the population size over time (and written to files) in the current simulation run to take into account reduced mortality in an intervention run with TB vaccines.
This correction requires the attribute value econ.output="true" in the XML inputfile (in <output><detailed.output>).

The optional argument VXa.incidence.files is a vector of filenames that will replace the file names in <VXa><VXa.incidence><incidence.data>. Note: only the filenames will be replaced, not the additional attributes. Use with care.

The directory structure of a specific country (ZAF as an example) and the required files are:
ZAF/data/demographics.csv 
ZAF/data/ZAF_deathrates.csv
ZAF/data/HIV-incidence.txt
ZAF/data/ART-incidence.txt
ZAF/data/ZAF_all_contacts_2020.txt
NOTE: the path and filename from the country folder depends on the specified path and filename in the XML input file e.g. <population file="data/demographics.csv"/>
ZAF/parameters/XMLinput.xml
ZAF/parameters/input.csv
ZAF/parameters/target.csv
NOTE: the filenames depend on the parameters passed to the intialize() function. The path is fixed.
ZAF/output	The folder where output will be written to
ZAF/logs       The folder where the model log will be written to
The initialize function also:
-	sets up the paths environment
-	sets up the global function modlog()
}
