% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{setup.model.log}
\alias{setup.model.log}
\title{Sets up the model log in the logs subfolder of the country folder as specified in the paths environment ; called from set.paths()}
\usage{
setup.model.log(paths, lglevel = "DEBUG")
}
\arguments{
\item{paths}{the paths environment from which the country folder will be derived}

\item{lglevel}{logging level}
}
\value{
a logging function to be assigned to the global function modlog() that will be called for logging e.g. modlog(level="FATAL", msg="not good')
}
\description{
Sets up the model log in the logs subfolder of the country folder as specified in the paths environment ; called from set.paths()
}
