% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{simulation.years}
\alias{simulation.years}
\title{reads the attributes from the XML element <simulation>
to be stored in params$simulation}
\usage{
simulation.years(xmldoc = NULL, h = NULL)
}
\arguments{
\item{xmldoc}{the XML document}

\item{h}{dt}
}
\value{
a named list of start, stop, and years (intermediate output moments) all in years
}
\description{
reads the attributes from the XML element <simulation>
to be stored in params$simulation
}
