% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-matrix-fncs-v12.R
\name{transition}
\alias{transition}
\title{Called from create.full.parameter.matrix.for.age.group()}
\usage{
transition(M = NULL, from = NA, to = NA, rate = NULL)
}
\arguments{
\item{M}{square matrix (e.g. transition / progression / treatment matrix)}

\item{from}{the name of the 'from' state}

\item{to}{the name of the 'to' state}

\item{rate}{the (additional) transition from M[from,from] to M[to,from] i.e. to be subtracted from M[from,from] and to be added to M[to,from]}
}
\value{
the modified matrix M
}
\description{
Called from create.full.parameter.matrix.for.age.group()
}
