% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-age-related-fncs-v14.R
\name{update.contact.matrix}
\alias{update.contact.matrix}
\title{Updates the contacts matrix to ensure that the number of contacts between A1 and A2 is equal to the number of contacts between A2 and A1}
\usage{
update.contact.matrix(parms, alive.pop)
}
\arguments{
\item{parms}{parameters environment}

\item{alive.pop}{current alive population size by age group}
}
\value{
updated contacts matrix
}
\description{
This function is called once per year at the start of the year in function aging.and.births.and.update.contact.matrix.event()
}
\details{
initializes VXa, SES, RISK, HIV, or TB progression from the XML specification
}
