% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-mod-xml-fncs-v2.R
\name{update.parameters}
\alias{update.parameters}
\title{reads the XML file and parses parameters unrelated to the model itself (i.e. dimensions, progression etc etc)}
\usage{
update.parameters(xmldoc, new.params = NULL)
}
\arguments{
\item{xmldoc}{the in memory XML document}

\item{new.params}{a data frame read from (usually) the 'input.csv' with parameter values to be replaced and to be located by the xpath in the .csv}
}
\value{
the modified XML document
}
\description{
reads the XML file and parses parameters unrelated to the model itself (i.e. dimensions, progression etc etc)
}
