% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{update.paths}
\alias{update.paths}
\title{update.paths() creates an R environment with file paths to the XML file, parameter file, targets file etc}
\usage{
update.paths(
  paths = NULL,
  mydir = here(),
  countries = "countries",
  countrycode = NA,
  targets = NA,
  xml = NA,
  baseline = NA,
  parameters = NA,
  VXa.incidence.files = NA
)
}
\arguments{
\item{paths}{the paths environment to be updated}

\item{mydir}{the ‘root’ directory}

\item{countries}{the name of the countries folder (which will be appended to mydir)}

\item{countrycode}{3 letter ISO code of the country (e.g. ZAF)}

\item{targets}{the filename of the targets.csv file}

\item{xml}{the filename of the XML input file}

\item{baseline}{the filename (without extent) of the XML file of the baseline}

\item{parameters}{the filename of the parameters file (which may overwrite parameters in the xml inputfile)}

\item{VXa.incidence.files}{a vector of filenames (including path from the country folder) to replace the VXa incidence files specified in the XMLinput.xml file}
}
\value{
the updated paths environment
}
\description{
For details see the set.paths() function that calls update.paths
}
