% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-query-fncs-v5.R
\name{validate.target.states}
\alias{validate.target.states}
\title{checks that target states are defined for dimensions}
\usage{
validate.target.states(target.states = NULL, p = NULL)
}
\arguments{
\item{target.states}{a named character vector of comma separated state name by dimension}

\item{p}{initialized model parameters}
}
\description{
checks that target states are defined for dimensions
}
