% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-parsing-xml-fncs-v13.R
\name{verify.data.frame.with.fixed.parameters}
\alias{verify.data.frame.with.fixed.parameters}
\title{verifies a data frame of fixed parameters for completeness of defined dependencies (i.e. all stages / age groups of a dimension)}
\usage{
verify.data.frame.with.fixed.parameters(p, z, xmlpath, age.groups = NULL)
}
\arguments{
\item{p}{the parameters environment}

\item{z}{the data frame to be verified}

\item{xmlpath}{the XML path of the parameters}

\item{age.groups}{the age groups defined}
}
\description{
verifies a data frame of fixed parameters for completeness of defined dependencies (i.e. all stages / age groups of a dimension)
}
