% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{write.econ.output}
\alias{write.econ.output}
\title{Writes econ output to files \code{###POP.txt}, \code{###dBGx.txt}, \code{###dHIVTBx.txt}, \code{###dPOPadj.txt}, \code{###dVx1pyr.txt}}
\usage{
write.econ.output(
  output = NULL,
  model.params = NULL,
  output.format = "txt",
  suffix = NA
)
}
\arguments{
\item{output}{output from the run() function including econ output (population, dBGx, dHIVTBx, dPOPadj, vaccinated)}

\item{model.params}{an environment with initialized model parameters}

\item{output.format}{the output format}
}
\description{
Writes econ output to files \code{###POP.txt}, \code{###dBGx.txt}, \code{###dHIVTBx.txt}, \code{###dPOPadj.txt}, \code{###dVx1pyr.txt}
}
