% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{write.merged.stocks.and.flows}
\alias{write.merged.stocks.and.flows}
\title{Writes merged stocks and flows output}
\usage{
write.merged.stocks.and.flows(
  onerun = NULL,
  model.params = NULL,
  output.format = "txt"
)
}
\arguments{
\item{onerun}{merged stocks and flows model output}

\item{model.params}{environment with initialized model parameters}

\item{output.format}{file format}
}
\description{
Writes merged stocks and flows output
}
