% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-output-fncs-v19.R
\name{write.output}
\alias{write.output}
\title{Writes model output to a file}
\usage{
write.output(data, filename, output.format = "txt")
}
\arguments{
\item{data}{data to be written (data.table or data.frame)}

\item{filename}{filename}

\item{output.format}{output format: either 'txt' (default) for tab delimited text, 
'parquet' for parquet and 'fst' for fst}
}
\value{
the modified list A after adding the result of fn to elements dY, dY.in and dY.out
}
\description{
Writes model output to a file
}
