% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{write.stocks.and.flows}
\alias{write.stocks.and.flows}
\title{Writes stocks and flows output}
\usage{
write.stocks.and.flows(
  output = NULL,
  model.params = NULL,
  output.format = "txt"
)
}
\arguments{
\item{output}{model output}

\item{model.params}{environment with initialized model parameters}

\item{output.format}{file format}
}
\description{
Writes stocks and flows output
}
