% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{write.targets}
\alias{write.targets}
\title{Writes the targets file with info on hit targets to a file}
\usage{
write.targets(output = NULL, model.params = NULL, output.format = "txt")
}
\arguments{
\item{output}{model output}

\item{model.params}{environment with initialized model parameters}

\item{output.format}{file format}
}
\description{
Writes the targets file with info on hit targets to a file
}
