% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-hi-level-fncs-v1.R
\name{write.updated.xml}
\alias{write.updated.xml}
\title{Writes the in memory XML doc that was updated with new parameter values to a file}
\usage{
write.updated.xml(parameters = NULL, xml = NULL)
}
\arguments{
\item{parameters}{environment with initialized (and updated) parameters}

\item{xml}{the new XML file name}
}
\description{
Writes the in memory XML doc that was updated with new parameter values to a file
}
