% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBVx-derivs-v27.R
\name{Y.matrix.from.X.vector}
\alias{Y.matrix.from.X.vector}
\title{Convert a vector of state variables to a matrix}
\usage{
Y.matrix.from.X.vector(X, p)
}
\arguments{
\item{X}{a double[] vector of state variables}

\item{p}{an environment with initialized model parameters}
}
\value{
a \code{double [p$nVXaSESRISKHIVTB x p$nAGES]} of state variables
}
\description{
\code{Y.matrix.from.X.vector()} converts a vector X of state variables to a matrix Y using parameters p
}
